server_stanczyk/requirements.txt
django==3.2.0
djangorestframework==3.11.1
django-ckeditor==6.3.1
django-filter==21.1
Pillow==9.1.0
